# apps/web/openapi/axon.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axon REST API",
    "description": "Dedicated REST routes for Axon discovery, RAG, crawl, ingest, and watch workflows.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "6.2.2"
  },
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "list_unified_jobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LifecycleStatus"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/JobKind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unified jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListPage"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "operationId": "clear_unified_jobs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobClearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unified job clear",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobClearResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/jobs/cleanup": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "cleanup_unified_jobs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCleanupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unified job cleanup",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCleanupResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/jobs/recover": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "recover_unified_jobs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRecoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unified job recovery",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecoveryResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "unified_job_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unified job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unified job status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/jobs/{id}/artifacts": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "unified_job_artifacts",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unified job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unified job artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobArtifactListResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/jobs/{id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "cancel_unified_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unified job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unified job cancellation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCancelResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/jobs/{id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "unified_job_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unified job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "since_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Severity"
            }
          },
          {
            "name": "visibility",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Visibility"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unified job event page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEventPage"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/jobs/{id}/retry": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "retry_unified_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unified job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unified job retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRetryResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "diagnostics": {
            "type": "object"
          },
          "kind": {
            "$ref": "#/components/schemas/ErrorKind"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "JobKind": {
        "type": "string",
        "enum": [
          "source",
          "watch",
          "map",
          "extract",
          "research",
          "ask",
          "query",
          "retrieve",
          "memory",
          "graph",
          "prune",
          "provider_probe",
          "reset"
        ]
      },
      "JobListPage": {
        "type": "object",
        "required": [
          "items",
          "limit"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "LifecycleStatus": {
        "type": "string",
        "enum": [
          "queued",
          "pending",
          "running",
          "waiting",
          "blocked",
          "canceling",
          "completed",
          "completed_degraded",
          "failed",
          "canceled",
          "expired",
          "skipped"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or static token"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/authorize",
            "tokenUrl": "/token",
            "scopes": {
              "axon:read": "Authenticated Axon REST access",
              "axon:write": "Authenticated Axon REST access"
            }
          }
        }
      }
    }
  },
...
}